Enabled webpack-dev-server
