+++ package.json
@@ -10,7 +10,8 @@
 		"start": "node dist/server.js",
 		"build-start": "npm run build && npm run start",
 		"test": "mocha --require ts-node/register 'test/**/*.ts'",
-		"cypress": "npx cypress open"
+		"cypress": "npx cypress open",
+		"start:dev": "webpack-dev-server --config webpack.dev.ts"
 	},
 	"dependencies": {
 		"@fortawesome/fontawesome-free": "^5.14.0",
@@ -68,6 +69,7 @@
 		"vue-template-compiler": "^2.6.11",
 		"webpack": "^4.44.1",
 		"webpack-cli": "^3.3.12",
+		"webpack-dev-server": "^3.11.0",
 		"webpack-merge": "^5.1.1",
 		"webpack-node-externals": "^2.5.1"
 	}
